Moves deleteMealItem and related functions into a separate file, adds listeners for all meal tables
